Make new file for collision/planning scene management
Don't try to move to a joint config if we're already very close to it
Add necessary Tahoma APIs to furnish joint state information
Fix some bugs introduced during refactoring
Remove pick status publishing while we find a nicer way to implement it
