+++ aurmr_tasks/src/aurmr_tasks/common/perception.py
@@ -15,8 +15,6 @@ from aurmr_perception.srv import (
 )
 from std_srvs.srv import Trigger
 
-from aurmr_perception.util import qv_mult, quat_msg_to_vec, vec_to_quat_msg
-
 from aurmr_tasks.util import add_offset
 
 class CaptureEmptyBin(State):
